Add WumpusController.
To test how this would all work, I created a new ZackScene that only
had my shit it in. If you load that up, you’ll see that there’s a
static “Wumpus/Spider” but there’s also a dynamic Wumpus/Spider that
moves around randomly and then dies!

ThingThatSpawnsWumpus is just my scene’s glue to make it happen - this
will be moved to DayPhase.